+++ src/util/mod.rs
@@ -2,4 +2,5 @@
 
 pub mod encoding;
 pub mod fs;
+pub mod postprocessors;
 pub mod scanner;